backend/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
dj-database-url==2.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.2
openai==1.55.3
langchain-core>=0.3.25
langchain==0.3.12
langchain-openai==0.2.12
langchain-community>=0.0.39
django-redis>=5.4.0
redis>=4.2.0
psycopg2-binary
gunicorn==21.2.0
yfinance==0.2.36
requests==2.31.0
beautifulsoup4==4.12.3
pandas>=2.0.0
numpy>=1.24.0
pinecone-client==3.0.2
langchain-pinecone>=0.0.3
tiktoken>=0.5.2
unstructured>=0.16.13
# Document loading dependencies
pdfminer.six>=20221105
python-docx>=0.8.11
markdown>=3.5.2
